[#12176] refactor(core): simplify policy tag associations
